Corrected failing tests (user creation did not include 'active')
When a user is created via these tests, the POST did not contain 'active:true' so the users could not login. When logins  fail the errors are a bit confusing. Also note that the Postman export decided to add 'body' objects with data instead of empty objects when exporting, so there are more diffs than were actually made. The fix here was to add 'active: true' to the POST of the user creation, which was in an env variable and several tests (4 places total).