Add 'active:true' to the createUser function
Argh! I *hate* this new version of Postman. The variable editor (stupidly) has an initial value and a current value. So, I edited the current value because it says the initial value would automatically be updated at the end of a test. Well, of course, it didn't even though the settings suggest it will and my first commit had the original variable, which I (again stupidly) did not check in my diff, due to Postman's crappy formatting and, as mentioned in the previous commit, decision to add bodies everywhere polluting my commit. I had to click persist all in the variable editor to update the initial value which actually causes the export to have the correct values.