Correct malformed JSON
The latest build of mod-users requires the "id" field to be an actual UUID, whereas older builds would take any string since this is ignored as it is written to the DB and a new UUID is returned. For some reason, "id" is a required field in the user schema which means that the client needs to send something and the schema does not enforce UUID format via a pattern. The documentation does indicate "id" must be a UUID. It appears that a UUID without the '-' characters is accepted, as the example in the API doc shows. Anyway, this test works now.