Fix mod-inventory-storage locations API tests
The locations managed by this API require a new property, `primaryServicePoint`. However, there is a runtime dependency on the new `servicePoints` array to contain the `primaryServicePoint`, so these needed to be added when using the locations API.

Also added `/service-points` API tests in order to have a service point for the locations API to use.