Correcting API Tests for mod-invoice
There is one voucher and voucher line which are part of sample data so these records are always available on reference environments. There are 2 requests which verify that there is no any voucher/line left in the storage once API Tests completed. These tests have to be removed because API Tests ideally should not rely on any initial state of the storage data:
- `Verify no Voucher Lines left in storage`
- `Verify no Voucher left in storage`